refactor(hronir): quebra commands.ts (3483 linhas, 20 exports) em commands/*
Move mecânica, zero mudança de comportamento: cada uma das 20 funções
exportadas de src/hronir/commands.ts foi movida para seu próprio arquivo em
src/hronir/commands/<nome-kebab-case>.ts (nome derivado do comando CLI real
em scripts/hronir/index.js). Helper privado usado por mais de um comando foi
para src/hronir/commands/_shared.ts, com `export` adicionado (antes eram
privados a um único módulo; agora precisam atravessar arquivo). Helper usado
por um único comando ficou junto dele no mesmo arquivo (ex.: fmt em
ranking.ts, parseRate/PROMPT_VERSION em decide.ts, collectDefenses/
getRecentlyEditedKeys/SKILLS_DIR em edit-worst.ts). src/hronir/commands/index.ts
é o barrel: reexporta as 20 funções com os mesmos nomes e assinaturas de hoje.

Compatibilidade de import: existem exatamente 2 call sites fora de
commands/ que importam de commands.js/commands.ts (scripts/hronir/index.js
e src/hronir/__tests__/flatten.test.js, este via import dinâmico com
cache-busting). Optei por manter um src/hronir/commands.ts fino
(`export * from "./commands/index.js"`) em vez de atualizar os 2 call
sites, porque isso resulta em zero arquivos externos tocados — o critério
de desempate era "menos arquivos mudados fora de commands/".

Verificação: extração feita por script (corte por range de linhas exatas,
não copy-paste manual), com checagem automatizada de que o multiset de
linhas não-vazias do arquivo original bate exatamente com a soma de todos
os arquivos gerados (zero conteúdo perdido/duplicado pela extração em si).
npm test: 78/78 verde, sem mudança nos arquivos de teste. npx astro check:
0 erros/0 warnings/73 hints, idêntico ao commands.ts original (comparado via
git stash). npx prettier --check: limpo. node scripts/check-hygiene.mjs:
limpo. npm run hronir:select && npm run hronir:doctor: "doctor: 0
inconsistências" (os 2 avisos não-bloqueantes são divergência de tradução
pré-existente, sem relação com este PR).

Bugs pré-existentes notados, não corrigidos (fora de escopo): PromoteArgs
(interface) e DiagnoseMatchEntry (interface) já estavam mortas — não usadas
em lugar nenhum do arquivo original — apenas movidas como estão (para
_shared.ts e diagnose.ts respectivamente).
