+++ src/hronir/commands/index.ts
@@ -1,0 +1,24 @@
+// Barrel re-exporting every Hrönir CLI command from its own file under
+// src/hronir/commands/. See src/hronir/commands.ts (the thin compat
+// shim at the pre-split import path) and scripts/hronir/index.js (the
+// CLI dispatcher that imports from here indirectly).
+export { init } from "./init.js";
+export { continueCmd } from "./continue.js";
+export { firstImpressionA } from "./first-impression-a.js";
+export { firstImpressionB } from "./first-impression-b.js";
+export { next } from "./next.js";
+export { generateMatch } from "./generate-match.js";
+export { getGlipho } from "./get-glipho.js";
+export { decide } from "./decide.js";
+export { submitEval } from "./submit-eval.js";
+export { ranking } from "./ranking.js";
+export { worst } from "./worst.js";
+export { diagnose } from "./diagnose.js";
+export { editWorst } from "./edit-worst.js";
+export { migrate } from "./migrate.js";
+export { doctor } from "./doctor.js";
+export { end } from "./end.js";
+export { editCommit } from "./edit-commit.js";
+export { select } from "./select.js";
+export { prune } from "./prune.js";
+export { flatten } from "./flatten.js";
